requirements.txt
numpy==1.23.4
Pillow==9.3.0
pypng==0.20220715.0
TBB==0.2
